Add SDK specific instruction
Using the Python SDK, there was an issue converting the protocol buffer response returned by the blockchain into a `dict` that can be serialized to JSON for the GetRolesPermissions and GetAllAccounts queries. Had to find the hard way that I needed a package to convert them into native datatypes.

Since it isn't part of the library, it might be worth pointing out to anyone who may need to use this command.